+++ corvid/containers/own_ptr.h
@@ -17,7 +17,7 @@
 
 #pragma once
 #include "containers_shared.h"
-#include "../infra/exception_wrappers.h"
+#include "../infra/exception_firewalls.h"
 
 #ifdef _WIN32
 #define NO_UNIQUE_ADDRESS [[msvc::no_unique_address]]
