+++ corvid/ecs/archetype_storage.h
@@ -27,7 +27,7 @@
 #include <utility>
 #include <vector>
 
-#include "../infra/exception_wrappers.h"
+#include "../infra/exception_firewalls.h"
 #include "archetype_storage_base.h"
 
 namespace corvid { inline namespace ecs { inline namespace archetype_storages {
