+++ corvid/ecs/entity_registry.h
@@ -26,7 +26,7 @@
 #include <type_traits>
 #include <utility>
 
-#include "../infra/exception_wrappers.h"
+#include "../infra/exception_firewalls.h"
 #include "../containers/fixed_bitset.h"
 #include "../meta/maybe.h"
 #include "entity_ids.h"
