+++ corvid/infra.h
@@ -16,5 +16,5 @@
 // limitations under the License.
 #pragma once
 #include "infra/clocks.h"
-#include "infra/exception_wrappers.h"
+#include "infra/exception_firewalls.h"
 #include "infra/log.h"
