+++ corvid/proto/epoll/epoll_stream_conn.h
@@ -33,7 +33,7 @@
 #include <unistd.h>
 #include <fcntl.h>
 
-#include "../../infra/exception_wrappers.h"
+#include "../../infra/exception_firewalls.h"
 #include "../iov_msghdr.h"
 #include "../net_endpoint.h"
 #include "../../concurrency/relaxed_atomic.h"
