+++ corvid/proto/io_uring/iou_dgram_router.h
@@ -23,7 +23,7 @@
 #include <utility>
 #include <vector>
 
-#include "../../infra/exception_wrappers.h"
+#include "../../infra/exception_firewalls.h"
 #include "../../containers/opt_find.h"
 #include "../../enums/bool_enums.h"
 #include "../net_endpoint.h"
